fix permissions issues

+++ apps/base/mixins/update.py
@@ -101,7 +101,23 @@ class UpdateMixin(utils.HelperMixin, AbstractUpdate):
         )
         
         if request.POST.get('delete'):
-            self.deleter(instance, request).delete()
+            perms = (
+                request
+                .user
+                .user_permissions
+                .filter(codename__startswith="delete_")
+            )
+            can_delete_models = [
+                p.content_type.model_class() for p in perms
+            ]
+            model = self._get_model_class()
+            if model in can_delete_models:
+                self.deleter(instance, request).delete()
+            else:
+                messages.error(
+                    request, 
+                    _('you can\'t delete this ({}) because you don\'t have permission to do so').format(instance)
+                )
             return self._get_success_save_update_response() 
         
         if form.is_valid():
